Config generator move file saver to its own class
